migrate http test utils to new http-client crate (#3888)
## Motivation and Context
<!--- Why is this change required? What problem does it solve? -->
<!--- If it fixes an open issue, please link to the issue here -->

Previous PR(s):
    * https://github.com/smithy-lang/smithy-rs/pull/3866

## Description
This PR migrates all of the HTTP related `test-util` feature to the new
`aws-smithy-http-client` crate.

* Re-export the relocated types from `aws-smithy-runtime`. I've tried to
ensure the feature gates line up correctly to what was there before. All
existing code depending on the `test-util` feature compiles and runs
still in `smithy-rs`.
* Updated the internals of all `test-util` code to be based on `hyper-1
/ http-1`.
* There were two functions `legacy_infallible::infallible_client_fn` and
`legacy_capture_request` that I had to just duplicate with the legacy
types to avoid a breaking change to them. These will be re-exported from
`aws-smithy-runtime` as the original types but otherwise not be used
going forward.

Next steps:
* Deprecate the old `test-util` APIs and update all of our usage of
`test-util` to be based on the new API's in `aws-smithy-http-client`. I
may ask for team to help in this effort as it should be relatively
straight forward migration work.

## Checklist
<!--- If a checkbox below is not applicable, then please DELETE it
rather than leaving it unchecked -->
- [x] For changes to the smithy-rs codegen or runtime crates, I have
created a changelog entry Markdown file in the `.changelog` directory,
specifying "client," "server," or both in the `applies_to` key.
- [x] For changes to the AWS SDK, generated SDK code, or SDK runtime
crates, I have created a changelog entry Markdown file in the
`.changelog` directory, specifying "aws-sdk-rust" in the `applies_to`
key.

----

_By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice._