chore: update url version to v2.5.4 (#3940)
## Motivation and Context
<!--- Why is this change required? What problem does it solve? -->
<!--- If it fixes an open issue, please link to the issue here -->

`idna` 0.5.0 which `url` depends on has CVE-2024-12224 so we need to
update the crate to v2.5.4.
https://rustsec.org/advisories/RUSTSEC-2024-0421

## Description
<!--- Describe your changes in detail -->

This PR is to update the crate on aws-config, aws-inlineable and
inlineable. Because all of them has dependency explicitly. Also I
updated Cargo.toml and run `cargo build` on each of them to generate
Cargo.lock.

## Testing
<!--- Please describe in detail how you tested your changes -->
<!--- Include details of your testing environment, and the tests you ran
to -->
<!--- see how your change affects other areas of the code, etc. -->

## Checklist
<!--- If a checkbox below is not applicable, then please DELETE it
rather than leaving it unchecked -->
- [ ] For changes to the smithy-rs codegen or runtime crates, I have
created a changelog entry Markdown file in the `.changelog` directory,
specifying "client," "server," or both in the `applies_to` key.
- [ ] For changes to the AWS SDK, generated SDK code, or SDK runtime
crates, I have created a changelog entry Markdown file in the
`.changelog` directory, specifying "aws-sdk-rust" in the `applies_to`
key.

----

_By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice._

---------

Signed-off-by: Ryota Sakamoto <skmt@amazon.com>
Co-authored-by: Landon James <lnj@amazon.com>