Fix serde on constrained blobs (#3893)
## Motivation and Context
#3890 

## Description
Replace the computed type (which will be a constrained shape, not what
we want!) with an absolute type.

## Testing
- [x] New unit test that fails without this fix applied.

## Checklist
<!--- If a checkbox below is not applicable, then please DELETE it
rather than leaving it unchecked -->
- [x] For changes to the smithy-rs codegen or runtime crates, I have
created a changelog entry Markdown file in the `.changelog` directory,
specifying "client," "server," or both in the `applies_to` key.


----

_By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice._