fix canary target to wasm32-wasip1 (#3966)
## Motivation and Context
<!--- Why is this change required? What problem does it solve? -->
<!--- If it fixes an open issue, please link to the issue here -->
`P180666804`


## Description
<!--- Describe your changes in detail -->

Release [1.84](https://releases.rs/docs/1.84.0/) removed support for the
`wasm32-wasi` target:

> Support for the target named wasm32-wasi has been removed as the
target is now named wasm32-wasip1. This completes the transition plan
for this target following the introduction of wasm32-wasip1 in Rust
1.78. Compiler warnings on use of wasm32-wasi introduced in Rust 1.81
are now gone as well as the target is removed.



Updates `wasm32-wasi` target to `wasm32-wasip1` which should fix the
canary on `stable` (example failed
[run](https://github.com/awslabs/aws-sdk-rust/actions/runs/12738274496)).


----

_By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice._