Fix flaky test in aws-sigv4 (#3990)
## Motivation and Context
The test was flaky because all invalid headers are counted as rejects
which can in rare cases exceed the maximum number of failures.

The test was flaky because all invalid headers are counted as rejects
which can in rare cases exceed the maximum number of failures.

## Description
Change the test to assert that the validity status matches instead of
filtering those cases out

## Testing
- cargo test


----

_By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice._

Co-authored-by: ysaito1001 <awsaito@amazon.com>