Add configuration ability to `mock_client` macro (#4001)
## Motivation and Context
We've seen folks who need to be able to modify configuration of the
client, but there's no way to deal with it outside of ejecting from the
macro.

## Description
- add configure option to the macro
- improve some docs

## Testing
I ran those doc tests unignored

## Checklist
<!--- If a checkbox below is not applicable, then please DELETE it
rather than leaving it unchecked -->
- [ ] For changes to the smithy-rs codegen or runtime crates, I have
created a changelog entry Markdown file in the `.changelog` directory,
specifying "client," "server," or both in the `applies_to` key.
- [ ] For changes to the AWS SDK, generated SDK code, or SDK runtime
crates, I have created a changelog entry Markdown file in the
`.changelog` directory, specifying "aws-sdk-rust" in the `applies_to`
key.

----

_By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice._