Merge documentation of same inline modules (#4008)
When two inline modules with the same name are generated from different
parts of the codebase, their documentation should be merged. However,
other metadata must match exactly, as it is an error for one part of the
codebase to define an inline module with pub visibility while another
defines it with pub(crate) visibility.

This PR enables documentation merging while maintaining strict
validation of other metadata fields.

Currently, the following sample model fails to generate code because
both `SomeList` and `member` are generated in the same inline module but
with different doc comments:

```smithy
@documentation("Outer constraint has some documentation")
@length(max: 3)
list SomeList {
    @length(max: 8000)
    member: String
}
```

Co-authored-by: Fahad Zubair <fahadzub@amazon.com>