Custom TLS config (#4032)
## Motivation and Context
<!--- Why is this change required? What problem does it solve? -->
<!--- If it fixes an open issue, please link to the issue here -->
* https://github.com/smithy-lang/smithy-rs/issues/1925
* https://github.com/awslabs/aws-sdk-rust/issues/977

## Description
Introduces a new `TlsContext` that can be used to customize TLS
configuration. Currently that only includes supplying custom root
certificates. This PR also refactors the way caching works. Previously
we cached complete connectors with default resolver when the connector
settings were known to be defaulted. The reason given in code and past
commits for this caching resolves around the fact that `hyper-rustls`
`with_native_roots()` configuration option re-loads platform root
certificates on each invocation. Rather than cache the connector,
instead I've switched the caching to only cache loading the
certificates. This reduces a bit of complexity and required context to
understand when we can cache.

### Questions/Discussion

1. There are a number of places we call `unwrap`/`expect` in this code.
It was this way previously but I've introduced a few more. This stems
from building a connector not returning a result. I don't know that I
love this but I also suppose I can understand why we chose this. Now is
the time to consider whether we'd approach it differently or not though
or we are simply saying a valid HTTPS configuration.
2. Loading native certs is all blocking file calls though and not async
friendly. I haven't changed anything here but it caught my eye. It would
be nice if this were more async friendly. The mitigation we chose to
this seemed to be caching the connector and paying that cost once up
front at construction time if possible.
3. Both rustls and s2n take in PEM encoded certs. Rustls (via
`rustls-pki-types`) provides a way to actually parse and load the certs
as an independent configuration object (trust store). The issue is it
converts them to DER format which makes using this for s2n not possible.
It would be nice if we could load and validate certs before passing them
off to rustls/s2n in a uniform way.
4. Should `TlsContext` and/or `TrustStore` provider builders? Currently
they just self mutate which is a pattern we use in some places (e.g.
stalled stream protection config comes to mind).


## Testing
New unit tests + existing TLS tests

----

_By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice._