fix event stream signing to include resource operations (#4055)
## Motivation and Context
fixes https://github.com/smithy-lang/smithy-rs/issues/4054

## Description
`serviceShape.operations` only includes operations directly bound to the
service and not operations bound via resources. This caused a bug with
event stream signing because the operation came from a resource and
didn't trigger the codegen path to enable the required feature flag in
the runtime (see the ticket for linked code). This PR address the
original issue and adds a test. I also searched for other places this
may be happening as it's a common bug in Smithy codegen.


## Checklist
<!--- If a checkbox below is not applicable, then please DELETE it
rather than leaving it unchecked -->
- [x] For changes to the smithy-rs codegen or runtime crates, I have
created a changelog entry Markdown file in the `.changelog` directory,
specifying "client," "server," or both in the `applies_to` key.

----

_By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice._